Extract payment setting from label settings into own component
